applied changes suggested by coderabbit

+++ frontend/components/Message.tsx
@@ -2,12 +2,14 @@
 import { useAuthStore } from "@/store/useAuthStore";
 import { deleteMessage, MessageType } from "@/utils/messages";
 import { flip, offset, shift, useFloating } from "@floating-ui/react";
+import { RefObject } from "react";
 
 interface Props {
     message: MessageType;
     menuIsOpen: boolean;
     selectedMessageID: number;
     handleRightClick: (messageID: number) => void;
+    containerRef: RefObject<HTMLDivElement | null>;
 }
 
 const Message = ({
@@ -15,6 +17,7 @@ const Message = ({
     menuIsOpen,
     selectedMessageID,
     handleRightClick,
+    containerRef,
 }: Props) => {
     const userID = useAuthStore((state) => state.userID);
     const date = new Date(message.CreatedAt);
@@ -33,13 +36,43 @@ const Message = ({
         placement: "right-start",
     });
 
-    const handleDelete = async () => {
+    const handleDelete = () => {
+        if (!confirm("Do you want delete this message? ")) return;
         deleteMessage(message.ID);
     };
 
+    const rect = containerRef.current?.getBoundingClientRect();
+    if (!rect) return;
+
     return (
         <div
+            tabIndex={0}
             className={`${message.SenderID === userID ? "bg-accent/80 ml-auto" : "bg-foreground/20 mr-auto"} py-[4px] px-[8px] rounded-[4px] w-fit flex flex-col`}
+            onKeyDown={(e) => {
+                if (e.key !== "Delete") return;
+                e.preventDefault();
+
+                if (message.SenderID !== userID) return;
+                handleRightClick(message.ID);
+                e.preventDefault();
+
+                refs.setPositionReference({
+                    getBoundingClientRect() {
+                        return {
+                            width: 0,
+                            height: 0,
+
+                            x: rect.left + rect.width / 2,
+                            y: rect.top + rect.height / 2,
+
+                            top: rect.top + rect.height / 2,
+                            left: rect.left + rect.width / 2,
+                            right: rect.left + rect.width / 2,
+                            bottom: rect.top + rect.height / 2,
+                        };
+                    },
+                });
+            }}
             onContextMenu={(e) => {
                 if (message.SenderID !== userID) return;
                 handleRightClick(message.ID);
@@ -76,7 +109,7 @@ const Message = ({
                 >
                     <button
                         aria-label="delete message"
-                        className="hover:bg-red-500 hover:text-foregrond cursor-pointer duration-300 rounded-[2px] p-[4px]"
+                        className="hover:bg-red-500 hover:text-foreground cursor-pointer duration-300 rounded-[2px] p-[4px]"
                         onClick={handleDelete}
                     >
                         Delete
